# Add cast benchmark.