Summary:
```
without error logging
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast_int##try_invalid                                        3.25s   307.91m
cast_int##try_valid                                         2.15ms    465.69
cast_int##valid                                             2.19ms    456.76
----------------------------------------------------------------------------
----------------------------------------------------------------------------
```
```
with error logging
============================================================================
[...]hmarks/ExpressionBenchmarkBuilder.cpp     relative  time/iter   iters/s
============================================================================
cast_int##try_invalid                                        4.88s   204.83m
cast_int##try_valid                                         2.15ms    465.22
cast_int##valid                                             2.20ms    453.96
----------------------------------------------------------------------------
----------------------------------------------------------------------------
```

Differential Revision: D47295115